Add icon SVG files for various technologies and LinkedIn
New SVG icons for Flask, Docker, Clojure and LinkedIn have been added to the 'public/icons' directory. These SVG files will be used to display corresponding tech logos on the website's tech stack section and the LinkedIn icon will be used in the footer for navigation to the LinkedIn page.